HomeTask3 MVC implemented
