+++ HomeTask3/index.html
@@ -2,14 +2,31 @@
     <head> 
         <meta name="viewport" content="width=device-width, initial-scale=1">
         <link rel="stylesheet" href="style.css">
+        <title>HomeTask3</title>
     </head>
     <body>
         <div class="header" id="header"></div>
         <div class="content" id="content"></div>
         <div class="footer" id="footer"></div>
-        <script language="javascript" type="text/javascript" src="content.js">
+        
+        
+        <script language="javascript" type="text/javascript" src="model/content.js">
         </script>
-        <script language="javascript" type="text/javascript" src="script.js">
+        <script language="javascript" type="text/javascript" src="view/modal.js" defer>
+        </script>
+        <script language="javascript" type="module" src="controller/funData.js" defer >
+        </script>
+        <script language="javascript" type="text/javascript" src="view/header.js" defer>
+        </script>
+        <script language="javascript" type="text/javascript" src="view/main.js" defer>
+        </script>
+        <script language="javascript" type="module" src="view/mainLeft.js" defer>
+        </script>
+        <script language="javascript" type="module" src="view/mainRight.js" defer>
+        </script>
+        <script language="javascript" type="text/javascript" src="view/footer.js" defer>
+        </script>
+        <script language="javascript" type="text/javascript" src="controller/funHandler.js" defer >
         </script>
     </body>
 </html>
