+++ HomeTask3/view/footer.js
@@ -1,0 +1,7 @@
+var foot = document.querySelector(".footer");
+
+var f = document.createElement("div");
+f.className = "footer"
+f.innerHTML = "&copy;&nbsp;NewsFeed 2019";
+f.style.textAlign = "center";
+foot.appendChild(f);
