+++ HomeTask3/view/main.js
@@ -1,0 +1,15 @@
+var main = document.querySelector(".content");
+
+//Main's Elements
+var c1 = document.createElement("div");
+c1.className = "c1";
+
+var c2 = document.createElement("div");
+c2.className = "c2";
+
+var c = document.createElement("div");
+c.className = "c";
+
+main.appendChild(c1);
+main.appendChild(c2);
+main.appendChild(c);
