+++ HomeTask3/view/modal.js
@@ -1,0 +1,12 @@
+var modal = document.createElement("div");
+document.querySelector('body').appendChild(modal);
+modal.className = "modal";
+modal.id = "myModal";
+var mdiv = document.createElement("div");
+mdiv.className = "modal-content";
+var mspan = document.createElement("span");
+mspan.id = "mspan";
+mspan.className = "close";
+mspan.innerHTML = "&times";
+modal.appendChild(mdiv);
+mdiv.appendChild(mspan);
